Clock time - usando HTML, CSS e JavaScript
Hey guys! I'm excited to share my new clock project that I developed using HTML, CSS and JavaScript. The goal was to create an interactive digital clock to display the current time.

Technologies used: HTML for structuring, CSS for styling and JavaScript for clock logic.
Next steps: In the future, I plan to add a stopwatch and alarm system to the watch to make it even more functional and useful.